+++ inc/core/patterns/cleanbc/goelectric-banner.php
@@ -1,0 +1,37 @@
+<?php
+/**
+ * CleanBC Banner block pattern.
+ *
+ * @since 1.0.0
+ *
+ * @package Bcgov/Theme/Block
+ */
+return array(
+	'title'      => __( 'CleanBC Go Electric Cards', 'bcgov_blocks_theme' ),
+	'categories' => array( 'bcgov-blocks-theme-patterns-cleanbc' ),
+	'content'    => '<!-- wp:cover {"url":"https://bcgov-block-theme.local/wp-content/uploads/2022/07/goElectricBC-shareable-01.jpg","id":306,"dimRatio":0,"customOverlayColor":"#000000","focalPoint":{"x":"0.49","y":"0.50"},"minHeight":50,"minHeightUnit":"vh","contentPosition":"center center","isDark":false,"align":"full"} -->
+<div class="wp-block-cover alignfull is-light" style="min-height:50vh"><span aria-hidden="true" class="wp-block-cover__background has-background-dim-0 has-background-dim" style="background-color:#000000"></span><img class="wp-block-cover__image-background wp-image-306" alt="" src="https://bcgov-block-theme.local/wp-content/uploads/2022/07/goElectricBC-shareable-01.jpg" style="object-position:49% 50%" data-object-fit="cover" data-object-position="49% 50%"/><div class="wp-block-cover__inner-container"><!-- wp:group {"layout":{"inherit":true}} -->
+<div class="wp-block-group"><!-- wp:columns {"align":"wide"} -->
+<div class="wp-block-columns alignwide"><!-- wp:column {"width":"30%","style":{"spacing":{"padding":{"top":"2rem","right":"2rem","bottom":"2rem","left":"2rem"}},"color":{"background":"#003366d1"}}} -->
+<div class="wp-block-column has-background" style="background-color:#003366d1;padding-top:2rem;padding-right:2rem;padding-bottom:2rem;padding-left:2rem;flex-basis:30%"><!-- wp:heading {"align":"wide","style":{"typography":{"fontSize":"3rem","lineHeight":"1.2"}},"className":"alignwide has-white-color has-text-color"} -->
+<h2 class="alignwide has-white-color has-text-color" style="font-size:3rem;line-height:1.2">More and more people are<strong><em> </em>going electric</strong></h2>
+<!-- /wp:heading -->
+
+<!-- wp:paragraph {"style":{"color":{"text":"#ffffff"}},"fontSize":"normal"} -->
+<p class="has-text-color has-normal-font-size" style="color:#ffffff">Through our CleanBC plan, the Province is making clean transportation more affordable and accessible for British Columbians.</p>
+<!-- /wp:paragraph -->
+
+<!-- wp:buttons -->
+<div class="wp-block-buttons"><!-- wp:button {"backgroundColor":"primary-brand-yellow","textColor":"primary-brand-blue","style":{"border":{"radius":"3px"}},"className":"is-style-fill"} -->
+<div class="wp-block-button is-style-fill"><a class="wp-block-button__link has-primary-brand-blue-color has-primary-brand-yellow-background-color has-text-color has-background" style="border-radius:3px">Learn More at Go Electric BC</a></div>
+<!-- /wp:button --></div>
+<!-- /wp:buttons --></div>
+<!-- /wp:column -->
+
+<!-- wp:column -->
+<div class="wp-block-column"></div>
+<!-- /wp:column --></div>
+<!-- /wp:columns --></div>
+<!-- /wp:group --></div></div>
+<!-- /wp:cover -->',
+);