+++ inc/core/theme-bootstrap.php
@@ -10,5 +10,8 @@
 // Add BC Gov Blocks Theme information page.
 require 'theme-info.php';
 
+// Add BC Gov Blocks Theme block filters.
+require 'filters/button-enhanced.php';
+
 // Add BC Gov Blocks Theme block patterns.
 require 'theme-block-patterns.php';